+++ next.config.js
@@ -1,14 +1,7 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const million = require('million/compiler')
-
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
 }
 
-const millionConfig = {
-  auto: true,
-}
-
-module.exports = million.next(nextConfig, millionConfig)
+module.exports = nextConfig
