+++ package.json
@@ -9,25 +9,23 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.3.1",
+    "@hookform/resolvers": "^3.1.1",
     "@phosphor-icons/react": "^2.0.10",
     "@types/node": "18.16.3",
     "@types/react": "18.2.0",
     "@types/react-dom": "18.2.1",
     "autoprefixer": "10.4.14",
     "daisyui": "^2.52.0",
-    "firebase": "^10.3.1",
-    "lodash": "^4.17.21",
-    "million": "2.6.0-beta.8",
+    "firebase": "^10.2.0",
     "next": "13.3.4",
     "next-seo": "^6.1.0",
-    "nodemailer": "^6.9.4",
+    "nodemailer": "^6.9.3",
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-awesome-slider": "^4.1.0",
     "react-dom": "18.2.0",
-    "react-hook-form": "^7.45.4",
-    "react-icons": "^4.10.1",
+    "react-hook-form": "^7.45.0",
+    "react-icons": "^4.9.0",
     "react-spinners": "^0.13.8",
     "sharp": "^0.32.5",
     "styled-components": "6.0.0-rc.1",
@@ -36,13 +34,12 @@
     "yup": "^1.2.0"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.197",
-    "@types/nodemailer": "^6.4.9",
-    "@typescript-eslint/eslint-plugin": "^5.62.0",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@types/nodemailer": "^6.4.8",
+    "@typescript-eslint/eslint-plugin": "^5.60.0",
+    "@typescript-eslint/parser": "^5.60.0",
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
-    "eslint-config-prettier": "^8.10.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^4.2.1",
     "prettier": "^2.8.8"
