+++ src/components/Header/Header.tsx
@@ -1,19 +1,44 @@
+import { useState, useEffect } from 'react'
 import { IBM_Plex_Mono } from 'next/font/google'
 import Link from 'next/link'
 import Image from 'next/image'
 import profilePicture from '/public/static/me.webp'
-import { HeaderProps } from '@/types'
-import useHeader from './hooks'
-import { memo } from 'react'
 
 const ibm = IBM_Plex_Mono({ subsets: ['latin'], weight: ['400', '700'] })
 
-const Header = memo(({ openMenu, isBuilding, transparent }: HeaderProps) => {
-  const { renderMenu, hidden, renderOpenMenu } = useHeader({
-    isBuilding,
-    transparent,
-    openMenu,
-  })
+interface HeaderProps {
+  openMenu: () => void
+  isBuilding: boolean
+  transparent?: boolean
+}
+
+const Header = ({ openMenu, isBuilding, transparent }: HeaderProps) => {
+  const [hidden, setHidden] = useState(!isBuilding || !transparent)
+
+  useEffect(() => {
+    if (isBuilding || transparent) return
+    let timeout: NodeJS.Timeout
+
+    const handleScroll = () => {
+      setHidden(true)
+      clearTimeout(timeout)
+      timeout = setTimeout(() => setHidden(false), 300)
+    }
+
+    const handleMouseMove = () => {
+      setHidden(false)
+      clearTimeout(timeout)
+      timeout = setTimeout(() => setHidden(true), 3000)
+    }
+
+    window.addEventListener('scroll', handleScroll)
+    window.addEventListener('mousemove', handleMouseMove)
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll)
+      window.removeEventListener('mousemove', handleMouseMove)
+    }
+  }, [transparent, isBuilding])
 
   return (
     <nav
@@ -52,11 +77,41 @@ const Header = memo(({ openMenu, isBuilding, transparent }: HeaderProps) => {
             ></path>
           </svg>
         </label>
-        {renderOpenMenu()}
-        {renderMenu()}
+        <input
+          type="checkbox"
+          className="hidden"
+          id="menu-toggle"
+          onClick={openMenu}
+        />
+        {!isBuilding && (
+          <div className="hidden lg:flex lg:items-center lg:w-auto" id="menu">
+            <ul className="flex flex-col lg:flex-row list-none gap-4">
+              <li className="nav-item">
+                <Link href="/" className="nav-link">
+                  Home
+                </Link>
+              </li>
+              <li className="nav-item">
+                <Link href="/about" className="nav-link">
+                  About
+                </Link>
+              </li>
+              <li className="nav-item">
+                <Link href="/works" className="nav-link">
+                  Works
+                </Link>
+              </li>
+              <li className="nav-item">
+                <Link href="/contact" className="nav-link">
+                  Contact
+                </Link>
+              </li>
+            </ul>
+          </div>
+        )}
       </div>
     </nav>
   )
-})
+}
 
 export { Header }
