+++ src/components/ProjectsList/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ProjectsList'
+export * from './ProjectsList'
