+++ src/components/SkillsList/index.ts
@@ -1,1 +1,1 @@
-export { default } from './SkillsList'
+export * from './SkillsList'
