+++ src/components/TwoThirdsOneThird/TwoThirdsOneThird.tsx
@@ -15,6 +15,7 @@ const TwoThirdsOneThird = ({
   return (
     <div className={`flex flex-col ${flexDir} p-4 md:p-12 w-full`}>
       <div className="w-full md:w-1/3 flex justify-center items-center">
+        {/* Imagem responsiva */}
         <img
           src={imagePath}
           alt="One Third"
@@ -22,9 +23,12 @@ const TwoThirdsOneThird = ({
           loading="lazy"
         />
       </div>
-      <div className="w-full md:w-2/3 text-gray-900">{text}</div>
+      <div className="w-full md:w-2/3 text-gray-900">
+        {/* Texto a ser exibido */}
+        {text}
+      </div>
     </div>
   )
 }
 
-export default TwoThirdsOneThird
+export { TwoThirdsOneThird }
