+++ src/components/UnderContructionPage/UnderConstruction.tsx
@@ -15,4 +15,4 @@ const UnderConstruction = () => {
   )
 }
 
-export default UnderConstruction
+export { UnderConstruction }
