+++ src/components/index.ts
@@ -1,5 +1,11 @@
 export * from './Header'
 export * from './Footer'
 export * from './layout'
+export * from './SkillsList'
+export * from './TwoThirdsOneThird'
 export * from './MobileMenu'
 export * from './SEO'
+export * from './ProjectsList'
+export * from './TopHomePage'
+export * from './UnderContructionPage'
+export * from './forms'
