--- src/components/inputs/InputText/index.ts
@@ -1,1 +1,0 @@
-export { default } from './InputText'
