+++ src/constants/projects.ts
@@ -1,0 +1,19 @@
+import { Project } from '@/types'
+
+export const projects: Project[] = [
+  {
+    name: 'PROJECT 1',
+    image: '/static/developer.webp',
+    description: 'This is a description for project',
+  },
+  {
+    name: 'PROJECT 2',
+    image: '/static/meeting.webp',
+    description: 'This is a description for project',
+  },
+  {
+    name: 'PROJECT 3',
+    image: '/static/404.webp',
+    description: 'This is a description for project',
+  },
+]
