+++ src/pages/works/index.tsx
@@ -1,6 +1,9 @@
+import dynamic from 'next/dynamic'
 import { GetLayout } from '@/interfaces/global'
 import { Layout } from '@/components'
-import ProjectCarousel from '@/components/ProjectCarousel/ProjectCarousel'
+import { useEffect } from 'react'
+import { useRedirect } from '@/hooks'
+const DynamicCarousel = dynamic(() => import('@/components/ProjectCarousel'))
 
 const getLayout: GetLayout = (page: JSX.Element) => {
   return (
@@ -11,7 +14,13 @@ const getLayout: GetLayout = (page: JSX.Element) => {
 }
 
 const Works = () => {
-  return <ProjectCarousel />
+  const { redirectHome } = useRedirect()
+
+  useEffect(() => {
+    redirectHome()
+  }, [redirectHome])
+
+  return <DynamicCarousel />
 }
 
 Works.getLayout = getLayout
